Add files via upload
This file generates a graph showing the distribution of displacement for a selected ion or atom in an MD simulation. It is particularly useful when comparing systems with a bound ion (such as chloride) between a wild-type protein and its mutant. Using this notebook in my research, I observed that one protein exhibited a more loosely bound chloride ion as a result of the introduced mutations.